+++ personal-hub-server/ph-system/src/test/java/com/personalhub/system/service/impl/UserServiceImplPasswordTest.java
@@ -26,7 +26,7 @@ class UserServiceImplPasswordTest {
 
     @BeforeEach
     void setUp() {
-        userService = new UserServiceImpl(userMapper, passwordEncoder);
+        userService = new UserServiceImpl(userMapper, passwordEncoder, null);
     }
 
     @Test